Added quantity sets in Ifc 4_0 properties schema
Added quantity sets Ifc 4_0 properties schema - they are just part of ifc4x0_properties.json, the schema is the same.
Added 4_0 and 2_3 to single DocExtractor class in doc.py.
