+++ src/ifcopenshell-python/ifcopenshell/util/schema/ifc4x0_property_sets_site_domains.json
@@ -418,5 +418,98 @@
     "pset_wasteterminaltypewastetrap": "ifcplumbingfireprotectiondomain",
     "pset_windowcommon": "ifcsharedbldgelements",
     "pset_workcontrolcommon": "ifcprocessextension",
-    "pset_zonecommon": "ifcproductextension"
+    "pset_zonecommon": "ifcproductextension",
+    "qto_actuatorbasequantities": "ifcbuildingcontrolsdomain",
+    "qto_airterminalbasequantities": "ifchvacdomain",
+    "qto_airterminalboxtypebasequantities": "ifchvacdomain",
+    "qto_airtoairheatrecoverybasequantities": "ifchvacdomain",
+    "qto_alarmbasequantities": "ifcbuildingcontrolsdomain",
+    "qto_audiovisualappliancebasequantities": "ifcelectricaldomain",
+    "qto_beambasequantities": "ifcsharedbldgelements",
+    "qto_boilerbasequantities": "ifchvacdomain",
+    "qto_buildingbasequantities": "ifcproductextension",
+    "qto_buildingelementproxyquantities": "ifcsharedbldgelements",
+    "qto_buildingstoreybasequantities": "ifcproductextension",
+    "qto_burnerbasequantities": "ifchvacdomain",
+    "qto_cablecarrierfittingbasequantities": "ifcelectricaldomain",
+    "qto_cablecarriersegmentbasequantities": "ifcelectricaldomain",
+    "qto_cablefittingbasequantities": "ifcelectricaldomain",
+    "qto_cablesegmentbasequantities": "ifcelectricaldomain",
+    "qto_chillerbasequantities": "ifchvacdomain",
+    "qto_chimneybasequantities": "ifcsharedbldgelements",
+    "qto_coilbasequantities": "ifchvacdomain",
+    "qto_columnbasequantities": "ifcsharedbldgelements",
+    "qto_communicationsappliancebasequantities": "ifcelectricaldomain",
+    "qto_compressorbasequantities": "ifchvacdomain",
+    "qto_condenserbasequantities": "ifchvacdomain",
+    "qto_constructionequipmentresourcebasequantities": "ifcconstructionmgmtdomain",
+    "qto_constructionmaterialresourcebasequantities": "ifcconstructionmgmtdomain",
+    "qto_controllerbasequantities": "ifcbuildingcontrolsdomain",
+    "qto_cooledbeambasequantities": "ifchvacdomain",
+    "qto_coolingtowerbasequantities": "ifchvacdomain",
+    "qto_coveringbasequantities": "ifcsharedbldgelements",
+    "qto_curtainwallquantities": "ifcsharedbldgelements",
+    "qto_damperbasequantities": "ifchvacdomain",
+    "qto_distributionchamberelementbasequantities": "ifcsharedbldgserviceelements",
+    "qto_doorbasequantities": "ifcsharedbldgelements",
+    "qto_ductfittingbasequantities": "ifchvacdomain",
+    "qto_ductsegmentbasequantities": "ifchvacdomain",
+    "qto_ductsilencerbasequantities": "ifchvacdomain",
+    "qto_electricappliancebasequantities": "ifcelectricaldomain",
+    "qto_electricdistributionboardbasequantities": "ifcelectricaldomain",
+    "qto_electricflowstoragedevicebasequantities": "ifcelectricaldomain",
+    "qto_electricgeneratorbasequantities": "ifcelectricaldomain",
+    "qto_electricmotorbasequantities": "ifcelectricaldomain",
+    "qto_electrictimecontrolbasequantities": "ifcelectricaldomain",
+    "qto_evaporativecoolerbasequantities": "ifchvacdomain",
+    "qto_evaporatorbasequantities": "ifchvacdomain",
+    "qto_fanbasequantities": "ifchvacdomain",
+    "qto_filterbasequantities": "ifchvacdomain",
+    "qto_firesuppressionterminalbasequantities": "ifcplumbingfireprotectiondomain",
+    "qto_flowinstrumentbasequantities": "ifcbuildingcontrolsdomain",
+    "qto_flowmeterbasequantities": "ifchvacdomain",
+    "qto_footingbasequantities": "ifcstructuralelementsdomain",
+    "qto_heatexchangerbasequantities": "ifchvacdomain",
+    "qto_humidifierbasequantities": "ifchvacdomain",
+    "qto_interceptorbasequantities": "ifcplumbingfireprotectiondomain",
+    "qto_junctionboxbasequantities": "ifcelectricaldomain",
+    "qto_laborresourcebasequantities": "ifcconstructionmgmtdomain",
+    "qto_lampbasequantities": "ifcelectricaldomain",
+    "qto_lightfixturebasequantities": "ifcelectricaldomain",
+    "qto_memberbasequantities": "ifcsharedbldgelements",
+    "qto_motorconnectionbasequantities": "ifcelectricaldomain",
+    "qto_openingelementbasequantities": "ifcproductextension",
+    "qto_outletbasequantities": "ifcelectricaldomain",
+    "qto_pilebasequantities": "ifcstructuralelementsdomain",
+    "qto_pipefittingbasequantities": "ifchvacdomain",
+    "qto_pipesegmentbasequantities": "ifchvacdomain",
+    "qto_platebasequantities": "ifcsharedbldgelements",
+    "qto_projectionelementbasequantities": "ifcproductextension",
+    "qto_protectivedevicebasequantities": "ifcelectricaldomain",
+    "qto_protectivedevicetrippingunitbasequantities": "ifcelectricaldomain",
+    "qto_pumpbasequantities": "ifchvacdomain",
+    "qto_railingbasequantities": "ifcsharedbldgelements",
+    "qto_rampflightbasequantities": "ifcsharedbldgelements",
+    "qto_reinforcingelementbasequantities": "ifcstructuralelementsdomain",
+    "qto_roofbasequantities": "ifcsharedbldgelements",
+    "qto_sanitaryterminalbasequantities": "ifcplumbingfireprotectiondomain",
+    "qto_sensorbasequantities": "ifcbuildingcontrolsdomain",
+    "qto_sitebasequantities": "ifcproductextension",
+    "qto_slabbasequantities": "ifcsharedbldgelements",
+    "qto_solardevicebasequantities": "ifcelectricaldomain",
+    "qto_spacebasequantities": "ifcproductextension",
+    "qto_spaceheaterbasequantities": "ifchvacdomain",
+    "qto_stackterminalbasequantities": "ifcplumbingfireprotectiondomain",
+    "qto_stairflightbasequantities": "ifcsharedbldgelements",
+    "qto_switchingdevicebasequantities": "ifcelectricaldomain",
+    "qto_tankbasequantities": "ifchvacdomain",
+    "qto_transformerbasequantities": "ifcelectricaldomain",
+    "qto_tubebundlebasequantities": "ifchvacdomain",
+    "qto_unitarycontrolelementbasequantities": "ifcbuildingcontrolsdomain",
+    "qto_unitaryequipmentbasequantities": "ifchvacdomain",
+    "qto_valvebasequantities": "ifchvacdomain",
+    "qto_vibrationisolatorbasequantities": "ifchvacdomain",
+    "qto_wallbasequantities": "ifcsharedbldgelements",
+    "qto_wasteterminalbasequantities": "ifcplumbingfireprotectiondomain",
+    "qto_windowbasequantities": "ifcsharedbldgelements"
 }
